refactor: consolidate UserInterest types into one file outside /me
